move taskDef so we can eliminate ParallelTasks
