+++ PYME/misc/taskDef.py
